requirements.txt
langgraph==0.2.14
langchain==0.2.15
langchain-groq==0.1.9
langchain-core==0.2.37
chromadb==0.5.5
streamlit==1.38.0
tavily-python==0.3.5
arxiv==2.1.3
sentence-transformers==3.0.1
networkx==3.3
plotly==5.24.1
python-dotenv==1.0.1
fpdf2==2.7.9
trafilatura==1.12.0
numpy==1.26.4
